third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-row-multi-line-block-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 1
  assert_equals: \n<item data-expected-margin-top="0" data-expected-margin-bottom="10" data-offset-y="8"></item>\noffsetTop expected 8 but got 18
[FAIL] flexbox > item 2
  assert_equals: \n<item data-expected-margin-top="0" data-expected-margin-bottom="10" data-offset-y="8"></item>\noffsetTop expected 8 but got 18
[FAIL] flexbox > item 3
  assert_equals: \n<item data-expected-margin-top="10" data-expected-margin-bottom="0" data-offset-y="78"></item>\noffsetTop expected 78 but got 88
[FAIL] flexbox > item 4
  assert_equals: \n<item data-expected-margin-top="10" data-expected-margin-bottom="0" data-offset-y="78"></item>\noffsetTop expected 78 but got 88
Harness: the test ran to completion.
